Fixing regression found in RCPTT tests
Class paths for jython were also passed to server for standard
libraries. This caused the jar files provided in class paths to be
locked even when there were only std libs used which resulted in not
being able to remove the project using it under windows due to locked
handle to a file.