RED-1336 Fixing failing JUnit after updating to RF 3.2
This test was RF version dependent because the construction:

Test Template  kw  1  2

is differently taken care of by different RF and thus our validation
also raises different errors depending on version. Prior to 3.2 the
arguments were joined together to form 'kw 1 2' and it was checked for
existence. From 3.2 onwards this just an error - there cannot be a
template having >1 arguments.

Change-Id: I10a4a110f4bc9e9a7344159be3959a532c82ccc2