RED-1432 Prepare for eclipse 2020-06
Fixing issue occuring in maven build when 4.16 platform is in use under
Java 8. It seems that starting from 4.16 they added javax.annotation
bundle to target platform but those packages where also available in
Java 8-10. This was causing some classes (e.g. @PostConstruct) to be
loaded twice with different class loaders. Then methods annotated with
them were not called because method injector was seeing two different
classes objects for the same class. In Java 11 javax.annotation was
removed so this problem was not visible.

After changing configuration of dependencies this should work in all
Java 8+

Change-Id: Ifaac5fa0deb74cdd4797d7e1115b7f57fc4a0aed